Update the way the calendar marks days
Now that we can adjust the calendar by year and month, having the list
of marked days be just days is problematic because the calendar can't
tell what dates the days are.

Here we modify the structure of the data so that it can convey the days
as dates in a minimal possible way (for use in the code block).

THe property is an array whose keys are years and whose values are
sub-objects, wherein the keys are months and the values are arrays of
the days in that month.

This allows for a robust, easy way to assign dates:

markedDays:
  2025:
    12: [6, 9]