Allow for multiple marker types in the calendar
The calendar now uses a different structure where the marked days can
now have a type in them:

markedDays:
  - type: default
    dates:
      2025:
        12: [6, 9]
  - type: darkslategrey
    dates:
      2025:
        12: [4, 20]

The type is either default to use the color variable, or is otherwise
used directly as the background color (so e.g. it could be a CSS color
or such).

This allows for multiple sets of marked days at once. The values are
added in the order given, so if the same day appears in more than one
list, they always appear in the color order given by the types.